Add non-grid pathfinding experiment
I couldn't find any pathfinding example for pico-8 that could be used for non-grid movement, so I took this one and build on top it: https://github.com/morgan3d/misc/blob/master/p8pathfinder/pathfinder.p8.

The necessity came when I started writing a game that had fluid movement inside a grid map (so rather than moving one tile at a time, the characters move one pixel at a time). The collision system was still grid-based, but the collisions had to be checked again at every pixel.

The strategy is:
1. Use A* at a tile grid level for long-distance pathfinding to find the shortest path to the target.
2. Get the closest tile to the character from the shortest path.
3. Use A* again at a pixel level to find the pixel path to that tile.

Of course it'd be possible to use A* at a pixel level for the whole path, but pico-8 limitations would be hit pretty quickly on a large map with many characters.